# backend/dev/OpenApi/openapi-site.yaml
openapi: 3.0.3
info:
  title: laravel-starter-kit
  version: 1.0.0
  description: 'Заготовка для проектов на Laravel с использованием Doctrine ORM, примеры модулей'
  contact:
    name: 'Студия 15'
    url: 'https://www.15web.ru'
    email: info@15web.ru
servers:
  -
    url: 'http://localhost:8088/api'
    description: 'Локальный сервер для разработки'
...
paths:
...
  /blog:
    get:
      operationId: indexBlog
      summary: 'Список записей блога'
      description: 'Список записей блога'
      tags:
        - blog
      security: []
      parameters:
        -
          $ref: '#/components/parameters/limit'
        -
          $ref: '#/components/parameters/offset'
      responses:
        '200':
          $ref: '#/components/responses/indexBlog'
    post:
      operationId: storeBlog
      summary: 'Создать запись в блоге'
      description: 'Создать запись в блоге'
      tags:
        - blog
      security: []
      requestBody:
        $ref: '#/components/requestBodies/storeBlog'
      responses:
        '200':
          $ref: '#/components/responses/storeBlog'
        '400':
          $ref: '#/components/responses/badRequest'
...
  /products/category:
    get:
      operationId: indexCategory
      summary: 'Дерево категорий товаров'
      description: 'Дерево категорий товаров'
      tags:
        - products
      security: []
      responses:
        '200':
          $ref: '#/components/responses/indexCategory'
    post:
      operationId: storeCategory
      summary: 'Создать категорию товаров'
      description: 'Создать категорию товаров'
      tags:
        - products
      security: []
      requestBody:
        $ref: '#/components/requestBodies/storeCategory'
      responses:
        '200':
          $ref: '#/components/responses/storeCategory'
        '400':
          $ref: '#/components/responses/badRequest'
components:
  securitySchemes:
    bearerHttpAuth:
      type: http
      scheme: bearer
      description: "Укажите токен без префикса `Bearer `, \\\nнапример '01928fb2-4147-7a36-a67b-88deaaee45b9'\n"
  headers:
    requestTraceId:
      schema:
        type: string
      description: 'Уникальный идентификатор запроса'
      example: 01904f4d-d18f-778e-a981-432d6af77425
...
  parameters:
    limit:
      in: query
      name: limit
      required: false
      description: Лимит
      example: 10
      schema:
        type: integer
    offset:
      in: query
      name: offset
      required: false
      description: Смещение
      example: 1
      schema:
        type: integer
...
  responses:
...
    badRequest:
      description: 'Ошибка запроса'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
          example:
            message: 'Неверный формат запроса'
            code: bad_request
            errors: []
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/requestTraceId'
...
    indexBlog:
      description: 'Возвращает список записей в блоге'
      content:
        application/json:
          schema:
            type: object
            required:
              - data
              - pagination
              - status
            additionalProperties: false
            properties:
              data:
                type: array
                items:
                  type: object
                  required: [id, title, author, createdAt]
                  additionalProperties: false
                  properties: { id: { $ref: '#/components/schemas/blogId' }, title: { type: string, description: 'Заголовок записи в блоге', minLength: 1, example: Введение }, author: { type: string, description: 'Автор блога', minLength: 1, example: Иван }, createdAt: { $ref: '#/components/schemas/createdAt' } }
              pagination:
                $ref: '#/components/schemas/pagination'
              status:
                $ref: '#/components/schemas/status'
          example:
            data:
              -
                id: 11
                title: Введение
                author: Иван
                createdAt: 2021-10-28T16:19:44+00:00
              -
                id: 12
                title: Заключение
                author: Потап
                createdAt: 2021-11-28T16:19:44+00:00
            pagination:
              total: 10
            status: success
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/requestTraceId'
    storeBlog:
      description: 'Запись в блоге создана'
      content:
        application/json:
          schema:
            oneOf:
              -
                type: object
                required:
                  - data
                  - status
                additionalProperties: false
                properties:
                  data: { type: object, required: [id, title, createdAt, updatedAt], additionalProperties: false, properties: { id: { $ref: '#/components/schemas/blogId' }, title: { type: string, description: 'Заголовок записи', minLength: 1, example: Введение }, createdAt: { $ref: '#/components/schemas/createdAt' }, updatedAt: { $ref: '#/components/schemas/updatedAt' } } }
                  status: { $ref: '#/components/schemas/status' }
              -
                $ref: '#/components/schemas/error'
          examples:
            'Успешный запрос':
              value:
                data:
                  id: 11
                  title: Введение
                  createdAt: 2024-10-15T10:59:14+00:00
                  updatedAt: null
                status: success
            'Запись с таким заголовком уже существует':
              value:
                message: 'Запись с таким заголовком уже существует'
                code: entity_exists
                errors: []
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/requestTraceId'
...
    indexCategory:
      description: 'Возвращает дерево категорий товаров'
      content:
        application/json:
          schema:
            type: object
            required:
              - data
              - pagination
              - status
            additionalProperties: false
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/productCategory'
              pagination:
                $ref: '#/components/schemas/pagination'
              status:
                $ref: '#/components/schemas/status'
          example:
            data:
              -
                id: 11
                title: Мебель
                children:
                  - { id: 13, title: Диваны, createdAt: 2021-11-28T16:19:44+00:00, updatedAt: null }
                  - { id: 14, title: Кровати, createdAt: 2021-11-28T16:19:44+00:00, updatedAt: null }
                createdAt: 2021-10-28T16:19:44+00:00
                updatedAt: null
              -
                id: 12
                title: Посуда
                children: []
                createdAt: 2021-11-28T16:19:44+00:00
                updatedAt: null
            pagination: null
            status: success
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/requestTraceId'
    storeCategory:
      description: 'Категория создана'
      content:
        application/json:
          schema:
            oneOf:
              -
                type: object
                required:
                  - data
                  - status
                additionalProperties: false
                properties:
                  data: { type: object, required: [id, title, createdAt, updatedAt], additionalProperties: false, properties: { id: { $ref: '#/components/schemas/categoryId' }, title: { type: string, description: 'Заголовок категории', minLength: 1, example: Диваны }, createdAt: { $ref: '#/components/schemas/createdAt' }, updatedAt: { $ref: '#/components/schemas/updatedAt' } } }
                  status: { $ref: '#/components/schemas/status' }
              -
                $ref: '#/components/schemas/error'
          examples:
            'Успешный запрос':
              value:
                data:
                  id: 11
                  title: Введение
                  createdAt: 2024-10-15T10:59:14+00:00
                  updatedAt: null
                status: success
            'Запись с таким заголовком уже существует':
              value:
                message: 'Запись с таким заголовком уже существует'
                code: entity_exists
                errors: []
            'Не найдена родительская категория':
              value:
                message: 'Не найдена родительская категория'
                code: bad_request
                errors: []
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/requestTraceId'
  schemas:
    status:
      type: string
      enum:
        - success
        - error
      description: 'Статус запроса'
      example: success
    error:
      type: object
      required:
        - message
        - code
        - errors
      additionalProperties: false
      properties:
        message:
          type: string
        code:
          type: string
          enum:
            - bad_request
            - unauthenticated
            - forbidden
            - not_found
            - method_not_allowed
            - unexpected
            - too_many_requests
            - entity_exists
        errors:
          type: array
          description: 'Список ошибок'
          items:
            type: string
            description: 'Текст ошибки'
            example: 'Cannot be empty'
    pagination:
      type: object
      nullable: true
      properties:
        total:
          type: integer
          description: 'Количество всех данных'
          example: 2
...
    createdAt:
      type: string
      description: 'Дата создания'
      example: '2021-10-28T16:19:44+00:00'
    updatedAt:
      type: string
      description: 'Дата обновления'
      example: '2021-10-28T16:19:44+00:00'
      nullable: true
...
    blogId:
      type: number
      minimum: 1
      description: 'Id записи'
      example: 11
...
    categoryId:
      type: number
      minimum: 1
      description: 'Id записи'
      example: 11
    productCategory:
      type: object
      required:
        - id
        - title
        - children
        - createdAt
        - updatedAt
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/categoryId'
        title:
          type: string
          description: 'Заголовок категории'
          minLength: 1
          example: Диваны
        children:
          type: array
          items:
            $ref: '#/components/schemas/productCategory'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        updatedAt:
          $ref: '#/components/schemas/updatedAt'
  requestBodies:
...
    storeBlog:
      description: 'Запрос на создание записи в блоге'
      content:
        application/json:
          schema:
            required:
              - title
              - author
              - content
            additionalProperties: false
            properties:
              title:
                type: string
                description: 'Заголовок записи в блоге'
                minLength: 1
                example: Введение
              author:
                type: string
                description: 'Автор блога'
                minLength: 1
                example: Иван
              content:
                type: string
                description: 'Текст записи'
                minLength: 1
                example: 'Начало рассказа...'
...
    storeCategory:
      description: 'Запрос на создание категории товаров'
      content:
        application/json:
          schema:
            required:
              - title
              - parent
            additionalProperties: false
            properties:
              title:
                type: string
                description: 'Заголовок категории'
                minLength: 1
                example: Диваны
              parent:
                type: number
                description: 'Id родительской категории'
                minimum: 1
                example: 11
                nullable: true